Remove single `.` from log message and fix one test
Most tooling (several IDEs and the macOS console) treat those dots as clickable part of the URL, which they are not. The commit changes that.

The commit also fixes a failing test (`io.quarkus.neo4j.deployment.Neo4jDevModeTests.DevServicesShouldBeAbleToUseFixedPorts`) that failed as the configuration for the properties under test had been renamed.